+++ src/components/shared/sidebar/index.ts
@@ -1,0 +1,2 @@
+export { ProjectTabToolbarSidebar } from './ProjectTabToolbarSidebar';
+export { SidebarButtonGroup } from './SidebarButtonGroup';